chore(admin): move maintenance actions into admin
Remove the standalone maintenance page and keep the remaining log cleanup actions behind the admin devices interface.

Made-with: Cursor
